CJLU_IMS-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncmy
aioredis
python-multipart
python-dotenv
alembic
pydantic
redis
pydantic-settings
PyJWT
cryptography
bcrypt
aiofiles
oss2
